refactor(web): split ChatPanel into composer hook + view components
Reduce components/ChatPanel.tsx from 791 to 304 lines, making it a layout
container. Behavior-preserving — verbatim logic/JSX moves.

- lib/hooks/useChatComposer.ts: controller hook owning the composer input/
  draft, slash-command menu, plan-mode toggle (+effects), file upload (wraps
  useFileUpload), refs, scroll bookkeeping, derived send-state flags, and the
  send/keyboard/slash handlers. Calls useModals/useGit internally and returns
  them so the view doesn't re-read the contexts.
- components/chat/ChatPanelSkeleton.tsx: dedupes the two near-identical
  loading/empty skeletons (parameterized title/header widths).
- components/chat/WelcomeView.tsx: new-chat welcome screen; takes the shared
  chatInput + filePreviewModal elements as props.
- components/chat/ChatMessageList.tsx: messages scroll region — bubbles,
  creating indicator, error/disconnected banners, queue shelf, scroll-to-bottom.

Typecheck clean (only the pre-existing Sidebar.tsx casing errors remain).
