+++ packages/web/components/chat/index.ts
@@ -4,3 +4,6 @@ export { PendingFilesDisplay } from "./PendingFilesDisplay"
 export { ChatHeader, MobileConflictBar } from "./ChatHeader"
 export { AgentModelSelector } from "./AgentModelSelector"
 export { ChatInput } from "./ChatInput"
+export { ChatPanelSkeleton } from "./ChatPanelSkeleton"
+export { WelcomeView } from "./WelcomeView"
+export { ChatMessageList } from "./ChatMessageList"
